requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
dnspython==2.3.0
fastapi==0.95.1
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gptcache==0.1.20
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
kiwisolver==1.4.4
langchain==0.0.142
llama-index==0.5.25
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
memory-profiler==0.61.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pinecone-client==2.2.1
psutil==5.9.5
pydantic==1.10.7
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.5.4
regex==2023.3.23
requests==2.29.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
